[refactor][product] Kafka product topic 통합
